requirements.txt
asgiref==3.7.2
Django==5.0.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
geopy==2.4.1
PyJWT==2.8.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
